# src/API/BlazoredDreams.Core/BlazoredDreams.Application/Comments/Models/CommentDto.cs
namespace BlazoredDreams.Application.Comments.Models
{
	public class CommentDto
	{
		public int Id { get; set; }
		public string Username { get; set; }
		public string Content { get; set; }
		public string Date { get; set; }
	}
}