# src/UI/src/api/swagger.json
{
  "x-generator": "NSwag v12.1.0.0 (NJsonSchema v9.13.28.0 (Newtonsoft.Json v11.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/Users": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_SignIn",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "204": {
            "description": ""
          },
          "default": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/Posts": {
      "get": {
        "tags": [
          "Posts"
        ],
        "operationId": "Posts_GetAll",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "page",
            "in": "query",
            "format": "int32",
            "default": 1,
            "x-nullable": false
          }
        ],
        "responses": {
          "default": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/PageOfPostPreviewDto"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "operationId": "Posts_Create",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "postRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "default": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/Posts/{id}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "operationId": "Posts_GetById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/PostDto"
            }
          },
          "404": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "default": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "operationId": "Posts_Delete",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "tags": [
          "Posts"
        ],
        "operationId": "Posts_Update",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "postRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "default": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/Dreams": {
      "get": {
        "tags": [
          "Dreams"
        ],
        "operationId": "Dreams_GetAll",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "page",
            "in": "query",
            "format": "int32",
            "default": 1,
            "x-nullable": false
          }
        ],
        "responses": {
          "default": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DreamDto"
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "Dreams"
        ],
        "operationId": "Dreams_Create",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "dreamRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DreamRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "default": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/Dreams/{id}": {
      "get": {
        "tags": [
          "Dreams"
        ],
        "operationId": "Dreams_GetById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/DreamDto"
            }
          },
          "404": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "default": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Dreams"
        ],
        "operationId": "Dreams_Delete",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "tags": [
          "Dreams"
        ],
        "operationId": "Dreams_Update",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "dreamRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DreamRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "default": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/posts/{postId}/comment": {
      "get": {
        "tags": [
          "Comments"
        ],
        "operationId": "Comments_GetAll",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "postId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "default": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CommentDto"
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "Comments"
        ],
        "operationId": "Comments_Create",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "postId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "commentRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/posts/{postId}/comment/{id}": {
      "delete": {
        "tags": [
          "Comments"
        ],
        "operationId": "Comments_Delete",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "postId",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "tags": [
          "Comments"
        ],
        "operationId": "Comments_Update",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "commentRequest",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentRequest"
            },
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "postId",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "default": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    }
  },
  "definitions": {
    "ProblemDetails": {
      "type": "object",
      "description": "A machine-readable format for specifying errors in HTTP API responses based on https://tools.ietf.org/html/rfc7807.",
      "additionalProperties": {},
      "properties": {
        "type": {
          "type": "string",
          "description": "A URI reference [RFC3986] that identifies the problem type. This specification encourages that, when\ndereferenced, it provide human-readable documentation for the problem type\n(e.g., using HTML [W3C.REC-html5-20141028]).  When this member is not present, its value is assumed to be\n\"about:blank\"."
        },
        "title": {
          "type": "string",
          "description": "A short, human-readable summary of the problem type.It SHOULD NOT change from occurrence to occurrence\nof the problem, except for purposes of localization(e.g., using proactive content negotiation;\nsee[RFC7231], Section 3.4)."
        },
        "status": {
          "type": "integer",
          "description": "The HTTP status code([RFC7231], Section 6) generated by the origin server for this occurrence of the problem.",
          "format": "int32"
        },
        "detail": {
          "type": "string",
          "description": "A human-readable explanation specific to this occurrence of the problem."
        },
        "instance": {
          "type": "string",
          "description": "A URI reference that identifies the specific occurrence of the problem.It may or may not yield further information if dereferenced."
        }
      }
    },
    "PageOfPostPreviewDto": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "currentPage",
        "pageSize",
        "totalPages"
      ],
      "properties": {
        "currentPage": {
          "type": "integer",
          "format": "int32"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        },
        "records": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PostPreviewDto"
          }
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "PostPreviewDto": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id",
        "comments"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "comments": {
          "type": "integer",
          "format": "int32"
        },
        "excerpt": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "tag": {
          "type": "string"
        }
      }
    },
    "PostDto": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id",
        "comments"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "tag": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "comments": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "PostRequest": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "dreamId"
      ],
      "properties": {
        "title": {
          "type": "string"
        },
        "excerpt": {
          "type": "string"
        },
        "dreamId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "DreamDto": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "date": {
          "type": "string"
        }
      }
    },
    "DreamRequest": {
      "type": "object",
      "additionalProperties": false,
      "properties": {
        "content": {
          "type": "string"
        }
      }
    },
    "CommentDto": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "date": {
          "type": "string"
        }
      }
    },
    "CommentRequest": {
      "type": "object",
      "additionalProperties": false,
      "properties": {
        "content": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "description": "Type into the textbox: Bearer {your JWT token}.",
      "name": "Authorization",
      "in": "header"
    }
  }
}